Pull request #3; refactor: filling up film with geners delegated to filmDao; refactor test - separated by classes;
