+++ src/main/java/ru/yandex/practicum/filmorate/dao/GenreDao.java
@@ -1,22 +1,13 @@
 package ru.yandex.practicum.filmorate.dao;
 
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 
-import java.util.Collection;
+import java.util.Optional;
 import java.util.Set;
 
 
 public interface GenreDao {
-    void updateToFilm(Film film);
-
-    void setGenresToAllFilms(Collection<Film> filmCollection);
-
-    public Genre getById(int genreId);
-
-    public Set<Genre> getByFilmId(int filmId);
-
-    public void deleteFromFilm(int filmId);
+    public Optional<Genre> getById(int genreId);
 
     public Set<Genre> getAll();
 }
