+++ src/main/java/ru/yandex/practicum/filmorate/dao/MpaDao.java
@@ -2,10 +2,11 @@ package ru.yandex.practicum.filmorate.dao;
 
 import ru.yandex.practicum.filmorate.model.Mpa;
 
+import java.util.Optional;
 import java.util.Set;
 
 public interface MpaDao {
-    public Mpa getById(int mpaId);
+    public Optional<Mpa> getById(int mpaId);
 
     public Set<Mpa> getAll();
 }
