+++ src/main/java/ru/yandex/practicum/filmorate/dao/mapper/GenreRowMapper.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.dao.mapper;
+
+import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.filmorate.model.Genre;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class GenreRowMapper implements RowMapper<Genre> {
+    @Override
+    public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return new Genre(rs.getInt("genre_id"), rs.getString("genre_name"));
+    }
+}
