+++ src/main/java/ru/yandex/practicum/filmorate/service/MpaService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dao.MpaDao;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Mpa;
 
 import java.util.Comparator;
@@ -23,6 +24,7 @@ public class MpaService {
     }
 
     public Mpa getMpaByMpaId(int mpaId) {
-        return mpaDao.getById(mpaId);
+        return mpaDao.getById(mpaId)
+                .orElseThrow(() -> new NotFoundException("Mpa not found by id: " + mpaId));
     }
 }
